Move informe inicial to gestion academica module
