+++ backend-ecep/src/main/java/edu/ecep/base_app/gestionacademica/domain/InformeInicial.java
@@ -1,5 +1,7 @@
-package edu.ecep.base_app.vidaescolar.domain;
+package edu.ecep.base_app.gestionacademica.domain;
 
+import edu.ecep.base_app.shared.domain.BaseEntity;
+import edu.ecep.base_app.vidaescolar.domain.Matricula;
 import jakarta.persistence.*;
 
 import java.time.LocalDate;
